camada controller atualizada

+++ src/main/java/com/fiap/auroratrace/java/controller/CameraController.java
@@ -25,7 +25,7 @@ public class CameraController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Camera> buscar(@PathVariable Long id) {
+    public ResponseEntity<Camera> buscar(@PathVariable Integer id) {
         return ResponseEntity.ok(service.buscarPorId(id));
     }
 
@@ -35,7 +35,7 @@ public class CameraController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletar(@PathVariable Long id) {
+    public ResponseEntity<Void> deletar(@PathVariable Integer id) {
         service.deletar(id);
         return ResponseEntity.noContent().build();
     }
